requirements.txt
fastapi
uvicorn[standard]
requests
qdrant-client
python-dotenv